hw13 done

+++ src/app.js
@@ -9,26 +9,28 @@ import { apiService } from './services/apiService';
 import { fetchImages, deleteItem } from './services/buildMarkup';
 import { scrollPage } from './utils/scroll';
 
-function searchFromInput(event) {
-  event.preventDefault();
+const searchFromInput = event => {
+  if (event.keyCode === 13) {
+    event.preventDefault();
 
-  const input = event.target;
-  if (input.value === '') {
-    messageNeedInput();
-    return;
+    const input = event.target;
+    if (input.value === '') {
+      messageNeedInput();
+      return;
+    }
+
+    deleteItem();
+
+    // console.log(input);
+    apiService.resetPage();
+    apiService.searchQuery = input.value;
+    fetchImages();
   }
 
-  deleteItem();
-
-  // console.log(input);
-  apiService.resetPage();
-  apiService.searchQuery = input.value;
-  fetchImages();
-
   // input.value = '';
-}
+};
 
-function loadMoreBtnHandler() {
+const loadMoreBtnHandler = () => {
   const input = apiService.query;
 
   if (!input) {
@@ -37,16 +39,8 @@ function loadMoreBtnHandler() {
   }
   fetchImages();
 
-  setTimeout(() => {
-    scrollPage();
-    // let element = document.querySelectorAll('ul > li:last-child');
-    // // console.log(element[0]);
-    // element[0].scrollIntoView({
-    //   behavior: 'smooth',
-    //   block: 'end',
-    //   inline: 'nearest',
-    // });
-  }, 1500);
-}
-refs.searchForm.addEventListener('input', debounce(searchFromInput, 1000));
+  setTimeout(() => scrollPage(), 1500);
+};
+
+refs.searchForm.addEventListener('keyup', debounce(searchFromInput, 500));
 refs.loadMoreBtn.addEventListener('click', loadMoreBtnHandler);
